working json serde on confs
Former-commit-id: 9497a7e14f7382ac84444c4544c5e67cb3b727f1

+++ deeplearning4j-core/src/main/java/org/deeplearning4j/models/featuredetectors/rbm/RBM.java
@@ -144,7 +144,7 @@ public  class RBM extends BaseNeuralNetwork {
 
 
         int k = conf.getK();
-        float learningRate = conf.getLr();
+        double learningRate = conf.getLr();
         int iteration = params[params.length - 1] == null ? 0 : (int) params[params.length - 1];
 
         if(wAdaGrad != null)
